move props from chat to message-form container
