+++ bigbluebutton-html5/imports/ui/components/chat/component.jsx
@@ -21,18 +21,6 @@ const intlMessages = defineMessages({
     id: 'app.chat.hideChatLabel',
     description: 'aria-label for hiding chat button',
   },
-  singularTyping: {
-    id: 'app.chat.singularTyping',
-    description: 'used to indicate when 1 user is typing',
-  },
-  pluralTyping: {
-    id: 'app.chat.pluralTyping',
-    description: 'used to indicate when multiple user are typing',
-  },
-  severalPeople: {
-    id: 'app.chat.severalPeople',
-    description: 'displayed when 4 or more users are typing',
-  },
 });
 const Chat = (props) => {
   const {
@@ -46,10 +34,6 @@ const Chat = (props) => {
     intl,
     shortcuts,
     isMeteorConnected,
-    typingUsers,
-    currentUserId,
-    startUserTyping,
-    stopUserTyping,
     lastReadMessageTime,
     hasUnreadMessages,
     scrollPosition,
@@ -60,47 +44,6 @@ const Chat = (props) => {
 
   const HIDE_CHAT_AK = shortcuts.hidePrivateChat;
   const CLOSE_CHAT_AK = shortcuts.closePrivateChat;
-
-  let names = [];
-
-  names = typingUsers.map((user) => {
-    const currentChatPartner = chatID;
-    const { userId: typingUserId, isTypingTo, name } = user;
-    let userNameTyping = null;
-    userNameTyping = currentUserId !== typingUserId ? name : userNameTyping;
-    const isPrivateMsg = currentChatPartner !== isTypingTo;
-    if (isPrivateMsg) {
-      const isMsgParticipant = typingUserId === currentChatPartner && currentUserId === isTypingTo;
-      userNameTyping = isMsgParticipant ? name : null;
-    }
-    return userNameTyping;
-  }).filter(e => e);
-
-  const renderIsTypingString = () => {
-    if (names) {
-      const { length } = names;
-      const noTypers = length < 1;
-      const singleTyper = length === 1;
-      const multipleTypersShown = length > 1 && length <= 3;
-      if (noTypers) return null;
-
-      if (singleTyper) {
-        if (names[0].length < 20) {
-          return ` ${names[0]} ${intl.formatMessage(intlMessages.singularTyping)}`;
-        }
-        return (` ${names[0].slice(0, 20)}... ${intl.formatMessage(intlMessages.singularTyping)}`);
-      }
-
-      if (multipleTypersShown) {
-        const formattedNames = names.map((name) => {
-          if (name.length < 15) return ` ${name}`;
-          return ` ${name.slice(0, 15)}...`;
-        });
-        return (`${formattedNames} ${intl.formatMessage(intlMessages.pluralTyping)}`);
-      }
-      return (` ${intl.formatMessage(intlMessages.severalPeople)} ${intl.formatMessage(intlMessages.pluralTyping)}`);
-    }
-  };
 
   return (
     <div
@@ -165,9 +108,6 @@ const Chat = (props) => {
           chatName,
           minMessageLength,
           maxMessageLength,
-          renderIsTypingString,
-          startUserTyping,
-          stopUserTyping,
         }}
         chatId={chatID}
         chatTitle={title}
