+++ bigbluebutton-html5/imports/ui/components/chat/message-form/component.jsx
@@ -22,9 +22,11 @@ const propTypes = {
   connected: PropTypes.bool.isRequired,
   locked: PropTypes.bool.isRequired,
   partnerIsLoggedOut: PropTypes.bool.isRequired,
-  renderIsTypingString: PropTypes.func.isRequired,
   stopUserTyping: PropTypes.func.isRequired,
   startUserTyping: PropTypes.func.isRequired,
+  currentChatPartner: PropTypes.string.isRequired,
+  currentUserId: PropTypes.string.isRequired,
+  typingUsers: PropTypes.arrayOf(Object).isRequired,
 };
 
 const defaultProps = {
@@ -56,6 +58,18 @@ const messages = defineMessages({
   errorChatLocked: {
     id: 'app.chat.locked',
   },
+  singularTyping: {
+    id: 'app.chat.singularTyping',
+    description: 'used to indicate when 1 user is typing',
+  },
+  pluralTyping: {
+    id: 'app.chat.pluralTyping',
+    description: 'used to indicate when multiple user are typing',
+  },
+  severalPeople: {
+    id: 'app.chat.severalPeople',
+    description: 'displayed when 4 or more users are typing',
+  },
 });
 
 const CHAT_ENABLED = Meteor.settings.public.chat.enabled;
@@ -76,6 +90,7 @@ class MessageForm extends PureComponent {
     this.handleMessageKeyDown = this.handleMessageKeyDown.bind(this);
     this.handleSubmit = this.handleSubmit.bind(this);
     this.setMessageHint = this.setMessageHint.bind(this);
+    this.renderIsTypingString = this.renderIsTypingString.bind(this);
   }
 
   componentDidMount() {
@@ -257,6 +272,53 @@ class MessageForm extends PureComponent {
     );
   }
 
+  renderIsTypingString() {
+    const {
+      intl, typingUsers, currentUserId, currentChatPartner,
+    } = this.props;
+    let names = [];
+
+    names = typingUsers.map((user) => {
+      const { userId: typingUserId, isTypingTo, name } = user;
+      let userNameTyping = null;
+      userNameTyping = currentUserId !== typingUserId ? name : userNameTyping;
+      const isPrivateMsg = currentChatPartner !== isTypingTo;
+      if (isPrivateMsg) {
+        const isMsgParticipant = typingUserId === currentChatPartner
+          && currentUserId === isTypingTo;
+
+        userNameTyping = isMsgParticipant ? name : null;
+      }
+      return userNameTyping;
+    }).filter(e => e);
+
+    if (names) {
+      const { length } = names;
+      const noTypers = length < 1;
+      const singleTyper = length === 1;
+      const multipleTypersShown = length > 1 && length <= 3;
+      if (noTypers) return null;
+
+      if (singleTyper) {
+        if (names[0].length < 20) {
+          return ` ${names[0]} ${intl.formatMessage(messages.singularTyping)}`;
+        }
+        return (` ${names[0].slice(0, 20)}... ${intl.formatMessage(messages.singularTyping)}`);
+      }
+
+      if (multipleTypersShown) {
+        const formattedNames = names.map((name) => {
+          if (name.length < 15) return ` ${name}`;
+          return ` ${name.slice(0, 15)}...`;
+        });
+        return (`${formattedNames} ${intl.formatMessage(messages.pluralTyping)}`);
+      }
+      return (` ${intl.formatMessage(messages.severalPeople)} ${intl.formatMessage(messages.pluralTyping)}`);
+    }
+
+    return null;
+  }
+
   render() {
     const {
       intl,
@@ -265,7 +327,6 @@ class MessageForm extends PureComponent {
       disabled,
       className,
       chatAreaId,
-      renderIsTypingString,
     } = this.props;
 
     const { hasErrors, error, message } = this.state;
@@ -309,8 +370,11 @@ class MessageForm extends PureComponent {
         </div>
         <div className={error ? styles.error : styles.info}>
           <span>
-            <span>{error || renderIsTypingString()}</span>
-            {!error && renderIsTypingString() ? <span className={styles.connectingAnimation} /> : null}
+            <span>{error || this.renderIsTypingString()}</span>
+            {!error && this.renderIsTypingString()
+              ? <span className={styles.connectingAnimation} />
+              : null
+            }
           </span>
         </div>
       </form>
